# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Jet on Set API",
        "version": "0.1"
    },
    "paths": {
        "/api/search/flight": {
            "get": {
                "tags": [
                    "Flight"
                ],
                "summary": "Display list of flights",
                "description": "Flights Page",
                "operationId": "a7ff28b8338bf306a55d8b36108c9463",
                "parameters": [
                    {
                        "name": "startPointName",
                        "in": "query",
                        "description": "Start Airport",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "endPointName",
                        "in": "query",
                        "description": "End Airport",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "flightDate",
                        "in": "query",
                        "description": "Date of flight",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "passengers",
                        "in": "query",
                        "description": "Count of passengers",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "startPoint",
                        "in": "query",
                        "description": "Departure geoId",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "endPoint",
                        "in": "query",
                        "description": "Arrival geoId",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}